requirements.txt
gradio==3.45.2
langchain==0.0.306
matplotlib==3.8.0
numpy==1.25.2
pandas==2.1.1
python-dotenv==1.0.0
youtube_transcript_api==0.6.1
